requirements.txt
onnxruntime
pillow
streamlit==1.24.0
numpy